Introducing hyperHTML.frog() Helper
A 🐸 is a bridge between a document fragment
and its inevitably lost list of rendered nodes

var render = hyperHTML.frog();
render`
  <div>Hello Frog!</div>
`;

Every single invocation will return that div
or the list of elements it contained as Array.
This simplifies most task where hyperHTML
is used to create the node itself, instead of
populating an already known and bound one.

Updated examples too.