Introducing incremental updates to close https://github.com/WebReflection/hyperHTML/issues/3
Added test/incremental.html and a simple test/todo.html to verify no focus is lost when new items are added.

The incremental update is linear, first to last child, and not backward, which might be a nice addition in the future but so far I'm quite happy with the result.