hyperHTML.wire(any?, 'adopt')
In order to be able to adopt lists, a new wire type
is used so that the wire does not resolve immediatly
but only once the parent element resolves its content.

Similar for hyperHTML.adopt, adopted wires can represent
only one element as container.

Still under consideration the possibility to make
adoption fully compatible with bound or wired nodes.

This is, however, a step forward to that direction.