Simplified non special attributes removal through null or undefined.
After discussion in #87 turns out there are cases where attributes
are not meant to be on the DOM.

Passing null, or undefined, as attribute value, and for non special attributes,
since these are already covered as boolean or events or input and others,
will now result into a temporary attribute removal.

The exact same attribute node will be eventually reattached
whenever the value changes to a non null or undefined value.