while benchmarking v2 I've noticed that majinbuu might perform very badly with huge amounts of changes (usually over few thousands).
Even if not always a realistic scenario, hyperHTML should perform as fast as possible by default so that a simplified engine can scale better for poorly constructed pages.

This was the opportunity to both separate the render/diff engine from the core, but also improve upfront most common Web use cases when handling lists:

append only
prepend only
changes in the middle only
When these common cases are not covered through the fast path, the rendering engine is involved.

The current engine by default is fast and simple but nearly as smart as majinbuu.

For those caring about moving the least amount of node, keeping lists size less than 1K of items, majinbuu is still the best option but it's now a module a part called hyperhtml-majinbuu.