hyperHTML: DOMSplicer based edition
Having a swappable engine should allow everyone to create
their best-for-needs implemnetation and DOMSplicer instances
are there to help.

It is now possible to create an engine that accepts an utils object
with a splicer, a default engine to fallback with,
as well as info about childNodes and where nodes are related
(which is the hyperHTML comment node).

This version of hyperHTML might be slightly slower on raw
operations but it should not affect overall performance.

https://github.com/WebReflection/dom-splicer