Component.for(generic) and new Component(generic)
In order to improve the situation described in #216
the `Component.for(item)` now passes the `item` along
when the new component is created (only the first time).

This makes users able to define their components
in ways compatible with both patterns so that
`new Component(data)` and `Component.for(data)`
would pass data around when/if necessary,
leaving to developers the choice on how/if store
such data internally or not.