Fix #225 - Comp.for(any) now different from Other.for(any)
There was an assertion that was actually meant to fail
but probably a copy and paste of the previous line made it pass.

Every refrence passed through a component is associated to it.
If two different classes/components use `.for(obj,..)` there
is no reason it should return the same reference.

Now it's properly tested.

Extra Fix: The WeakMap poorlyfill was leaking instances.
Now it doesn't anynmore.