Introducing hyperHTML.define("hyper-attribute", callback)
As asked in #256 there is no easy way to hook your code into attributes.

This update introduce the ability to define attributes simply using
a naming convention where any `attribute-name` will be registered
and its callback invoked with the current node, as `target`, and
whatever value was passed along, which could be also null.