Introducing domtagger as external dependency
In an ongoing attempt to refactor out all the bricks that make
hyperHTML what it is, I have also removed all the unnecessary code
that is taken care now by domtagger, which is 100% code covered a part,
together with all other external dependencies.

This MR also cleans up the test and its coverage, hoping to see
the coveralls badge showing some number again.